- new include/snprintf.[c|h] (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@1396 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ include/functions.c
@@ -755,7 +755,7 @@ int	evaluate_FUNCTION_thread(MYSQL *database, char *value,DB_ITEM *item,char *fu
 			value_float=value_float/(1024*1024*1024);
 		}
 		zabbix_log( LOG_LEVEL_WARNING, "Value [%s] [%f] Suffix [%s] Units [%s]",value,value_float,suffix,item->units);
-		if(cmp_double(round(value_float), value_float) == 0)
+		if(cmp_double((double)round(value_float), value_float) == 0)
 		{
                 	snprintf(value, MAX_STRING_LEN-1, "%.0f %s%s", value_float, suffix, item->units);
 		}
@@ -1014,7 +1014,7 @@ int	evaluate_FUNCTION(char *value,DB_ITEM *item,char *function,char *parameter, 
 			value_float=value_float/(1024*1024*1024);
 		}
 		zabbix_log( LOG_LEVEL_WARNING, "Value [%s] [%f] Suffix [%s] Units [%s]",value,value_float,suffix,item->units);
-		if(cmp_double(round(value_float), value_float) == 0)
+		if(cmp_double((double)round(value_float), value_float) == 0)
 		{
                 	snprintf(value, MAX_STRING_LEN-1, "%.0f %s%s", value_float, suffix, item->units);
 		}
